test_requirements.txt
Pillow==3.0.0
Unidecode==0.04.18
argparse==1.2.1
beautifulsoup4==4.4.1
django-classy-tags==0.6.2
django-filer==1.0.2
django-mptt==0.7.4
django-polymorphic==0.7.2
django-sekizai==0.8.2
django-treebeard==3.0
djangocms-admin-style==1.0.5
djangocms-text-ckeditor==2.7.0
easy-thumbnails==2.2.1
html5lib==0.9999999
jsonfield==1.0.3
pluggy==0.3.1
py==1.4.30
pytest==2.8.2
pytest-django==2.9.1
six==1.10.0
tox==2.2.1
virtualenv==13.1.2
django-reversion==1.9.3
